refactor(modals): extract modal model + codec into activeModal.ts
Move the modal vocabulary and URL codec out of the generic store/actions.ts
into store/activeModal.ts: URL_MODAL_IDS, UrlModalIdSchema, MODAL_RENAMES,
ModalIdSchema, ModalId, ActiveModal, encodeActiveModal, modalOf, and decodeShow
renamed to decodeActiveModal (symmetric with encodeActiveModal).

setActiveModal stays in actions.ts (it's an action; imports the type), so its
callers are untouched. As a bonus the codec lives outside the module harvested
by rootAction.ts, so its string/object return types can't pollute RootAction.
