+++ src/shared/hooks/useMenuHandler.ts
@@ -1,8 +1,6 @@
 import {
   clearMapFeatures,
   ExternalTarget,
-  ModalId,
-  modalOf,
   openInExternalApp,
   saveSettings,
   setActiveModal,
@@ -11,6 +9,7 @@ import {
   Tool,
   ToolSchema,
 } from '@app/store/actions.js';
+import { ModalId, modalOf } from '@app/store/activeModal.js';
 import { Document, documentShow } from '@features/documents/model/actions.js';
 import { l10nSetChosenLanguage } from '@features/l10n/model/actions.js';
 import { Submenu } from '@features/mainMenu/components/submenu.js';
